fix: resolve all clippy warnings
- Add Default impl for OpenAIWsStateStore
- Replace Iterator::last with next_back
- Replace or_insert_with(Vec::new) with or_default()
- Remove unnecessary cast
- Replace parts.get(0) with parts.first()
- Collapse nested if statements
- Replace redundant closure with method reference
- Use strip_prefix instead of manual slicing
- Replace Vec::new() + push with vec![] macro
- Replace useless format! with to_string()
- Replace match with matches! macro
- Remove unnecessary map(|s| s)
- Add #[derive(Default)] for UsageLimits
- Fix needless borrows for generic args
- Replace vec![] with array where applicable
- Allow upper_case_acronyms for HTTP methods and storage types
- Replace map_or with is_some_and for clearer logic
- Use range contains for temperature/top_p validation

+++ backend/src/service/timing_wheel_service.rs
@@ -192,7 +192,7 @@ impl TimingWheelService {
         stats.completed_tasks += tasks_to_execute.len() as u64;
         stats.pending_tasks = stats
             .pending_tasks
-            .saturating_sub(tasks_to_execute.len() as usize);
+            .saturating_sub(tasks_to_execute.len());
         stats.current_tick = current_tick;
 
         tasks_to_execute
